ipsec whack: move whack_{each,all}_connections() to whack_connection.[hc]
